# cli: optimize reana-client