--- a/reana_client/cli/ping.py
+++ b/reana_client/cli/ping.py
@@ -11,7 +11,6 @@ import logging
 import traceback
 
 import click
-from reana_client.api.client import ping as rs_ping
 from reana_client.cli.utils import check_connection
 from reana_client.version import __version__
 
@@ -34,7 +33,8 @@ def ping(ctx):  # noqa: D301
     \t $ reana-client ping
     """
     try:
-        from reana_client.cli import get_api_url
+        from reana_client.api.client import ping as rs_ping
+        from reana_client.utils import get_api_url
         logging.info('Connecting to {0}'.format(get_api_url()))
         response = rs_ping()
         click.echo(click.style('Connected to {0} - Server is running.'