--- a/reana_client/cli/utils.py
+++ b/reana_client/cli/utils.py
@@ -15,7 +15,6 @@ import sys
 
 import click
 
-from reana_client.api.client import current_rs_api_client
 from reana_client.utils import workflow_uuid_or_name
 from reana_client.config import ERROR_MESSAGES
 from reana_commons.errors import MissingAPIClientConfiguration
@@ -48,13 +47,12 @@ def check_connection(func):
     """Check if connected to any REANA cluster."""
     @functools.wraps(func)
     def wrapper(*args, **kwargs):
-        try:
-            _url = current_rs_api_client.swagger_spec.api_url
-        except MissingAPIClientConfiguration as e:
+        from reana_client.cli import get_api_url
+        api_url = get_api_url()
+        if not api_url:
             click.secho(
                 'REANA client is not connected to any REANA cluster.',
-                fg='red', err=True
-            )
+                fg='red', err=True)
             sys.exit(1)
         return func(*args, **kwargs)
     return wrapper