closes #354 

Optimizations:
* `ping`: ~2.2s → ~0.8s
* `--help`: ~0.6s → ~0.3s
* `version`: ~0.6s → ~0.3s
* `create`: ~2s → ~1.2s
* `start`: ~2.4s → ~1.5s
* `status`: ~1.6s → ~1.2s
* `run`: ~7s → ~3.9s (reana-demo-worldpopulation)
* `list`: ~1.6s → ~1.2s
* `diff`: ~1.6s → ~1s
* `logs`: ~1.6s → ~1s
* `stop`: ~1s → ~0.4s
* `upload`: ~2.3s → ~1.2s
* `download`: ~1.8s → ~1.1s
* `du`: ~1.5s → ~0.9s
* `ls`: ~2s → ~1.6s
* `mv`: ~2.9s → ~2.3s
* `rm`: ~1.6s → ~1s