@tiborsimko 2020-01-21
Nice improvements!  And we have a room for more, e.g. compare `reana-client version` with `reana-dev version`:

```console
$ python -m cProfile -o rc.prof ~/.virtualenvs/reana/bin/reana-client version
$ echo -e 'sort cumtime\nstats' | python -m pstats rc.prof | less
         453917 function calls (443016 primitive calls) in 1.018 seconds
   ...
        1    0.000    0.000    0.506    0.506 /home/simko/private/project/reana/src/reana-client/reana_client/utils.py:8(<module>)
        1    0.000    0.000    0.500    0.500 /home/simko/.virtualenvs/reana/lib/python3.8/site-packages/yadageschemas/__init__.py:1(<module>)
        1    0.000    0.000    0.270    0.270 /home/simko/.virtualenvs/reana/lib/python3.8/site-packages/pkg_resources/__init__.py:2(<module>)
        1    0.000    0.000    0.236    0.236 /home/simko/.virtualenvs/reana/lib/python3.8/site-packages/jsonschema/__init__.py:1(<module>)
     2151    0.008    0.000    0.235    0.000 /home/simko/.virtualenvs/reana/lib/python3.8/re.py:287(_compile)
      172    0.000    0.000    0.226    0.001 /home/simko/.virtualenvs/reana/lib/python3.8/re.py:248(compile)
```

with ~0.5 seconds spent in yadage schemas and ~0.3 seconds with the regular expression business handling (2K compilations!), while `reana-dev` is ultra fast here:

```console
$ python -m cProfile -o rc.prof ~/.virtualenvs/reana/bin/reana-dev version
$ echo -e 'sort cumtime\nstats' | python -m pstats rc.prof | less
28090 function calls (27441 primitive calls) in 0.098 seconds
```

We can think of these further improvements later.