@diegodelemos 2020-01-21
All commands tested, everything working except `reana-client delete` (see [here](https://github.com/reanahub/reana-client/pull/359/files#r368997402)). Very smooth 🏄‍♂️